+++ src/peak2anno/__init__.py
@@ -1,4 +1,3 @@
 """Command-line tools for annotating genomic peaks."""
 
-__version__ = "0.1.0"
-
+__version__ = "0.1.7"